Use latest wasm-opt instead of one in wasm-pack
wasm-pack auto-downloads and runs a fixed version of wasm-opt, but wasm-pack itself hasn't been maintained for a while, and so wasm-opt that comes with it is also severely outdated, leading to all sorts of hidden issues (compiler errors, broken atomics code, etc.) as well as missed optimisations.

This change disable wasm-opt feature of wasm-pack in all Rust codecs and runs the latest wasm-opt manually instead.