Tidy up CLI implementation
A couple of issues while working on it:
1. The `processFiles` function is `async`. If it would throw an error,
then `commander` wouldn't properly catch the error, which results in
an uncaught promise rejection. Instead, we should be using `parseAsync`,
which can handle an async action
2. If the user would specify the encoder, but without encoding configuration,
then the CLI would crash. Now, double check that we the user passes in
a string and, if not, throw a nicer error.
3. For some reason, creating the `ora` spinner in the `progressTracker`
function results in leaked listeners when listening on `stdout`. Presumably
that's because `commander` already listens to it, but I don't know for sure.
If we instead call `ora` in the module scope, it can attach to the proper
outputstream and ensure that the leak is prevented.