Refactor CLI to perform batch processing
To avoid memory pressure by a complete sequential pipeline, instead
we can refactor the CLI to perform batch processing. For the number
of threads that Squoosh spawns, perform decoding and encoding in
sequence. Then, move on the next batch (if it exists).

I have confirmed that compressing 300 high resolution images works
with batch processing.

To make sure that the progress is appropriately shown, the
already-encoded images are now shown in the results, while the CLI
is doing its work. At the end, rather than printing all results
again, it only prints the total number of images. Otherwise, it
would result in a lot of duplicate console spam, which is not
desirable.

To make sure that all numbers are correct, the progress tracker is
refactored to instead use `progressOffset` to keep track of previous
batches. Additionally, rather than having the `total` set every
single time, it now sets it once at initialization (with the number
of files in total to compress)

Fixes #1012