Use automatic detection of environment for hardware concurrency
We can pass in any string into `PTHREAD_POOL_SIZE`, which means we can
inject a bit of code to implement the switch-case for both Node and web.
On the web, it will use `navigator.hardwareConcurrency` as before, while
on Node it will use `cpus().length`. This allows us to remove the
assignment to `navigator` on the global scope in `codecs.ts`, which
avoids throwing off tools like `sass` which assume a web environment
when that global is set.

Fixes #1152